Add functionality to visual query modal search button
