+++ src/features/entities/search-form.tsx
@@ -59,17 +59,6 @@ export function SearchForm(): JSX.Element {
       >
         Create visual query
       </Button>
-
-      <Button
-        type="button"
-        round="round"
-        color="accent"
-        size="regular"
-        className="col-span-2"
-        onClick={openVisualQueryingModal}
-      >
-        Create visual query
-      </Button>
     </form>
   );
 }
